requirements.txt
numpy
pandas
python-igraph
scipy
networkx
node2vec
xgboost
scikit-learn
